pyproject.toml
[tool.poetry]
name = "trezor-firmware"
version = "0.0.0"
description = ""
authors = ["SatoshiLabs <info@satoshilabs.com>"]

[tool.poetry.dependencies]
# all
python = "^3.6"
trezor = {path = "./python", develop = true}
scons = "*"
protobuf = "*"
pyblake2 = "*"
Pyro4 = "*"
nanopb = "^0.4.3"

## test tools
pytest = "^6.1.0"
coverage = "<5"  # coverage 5+ uses binary sqlite format
flaky = "^3.6.1"  # https://github.com/box/flaky/issues/156
pytest-ordering = "*"
pytest-random-order = "*"
pytest-timeout = "*"
tox = "*"
dominate = "*"

# hardware tests
pyserial = "*"

## test requirements
shamir-mnemonic = "^0.2.1"
fido2 = "^0.8.0"

## typing
mypy = "^0.910"
typing-extensions = ">=3.7.4"

## style
isort = "<5"  # 5 changes lots of stuff that need to be addressed first: https://timothycrosley.github.io/isort/docs/upgrade_guides/5.0.0/
flake8 = ">=3.7.0"
black = ">=20.8b0"
mako = "^1.0.7"
munch = "^2.3.2"
autoflake = "*"
flake8-requirements = ">=1.3.2"
pylint = "^2.11"

# common
demjson3 = "*"
graphviz = "*"

## cointool
click = "^7"
ed25519 = "^1.4"
requests = "^2.19"
termcolor = "*"
Pillow = "^8.0"

# crypto
ecdsa = "^0.16"
curve25519-donna = "*"
pyasn1 = "*"

# legacy
setuptools = ">=24.2.0"

# storage
cryptography = "*"
hypothesis = "*"
attrs = "*"
inotify = "*"
yamllint = "^1.25.0"

[tool.poetry.dev-dependencies]
scan-build = "*"
towncrier = "^21.3.0"

[build-system]
requires = ["poetry>=1;<1.1", "pip>=20"]
build-backend = "poetry.masonry.api"
